test: small refactoring
