+++ QA/py/data/import_issues/predicted_but_what/m106_mn01_n3_sml/ecotaxa_m106_mn01_n3_sml_pls.tsv
@@ -1,0 +1,3 @@
+object_id	object_extra	object_lat	object_lon	object_date	object_time	object_link	object_depth_min	object_depth_max	object_annotation_status	object_annotation_person_name	object_annotation_person_email	object_annotation_date	object_annotation_time	img_file_name	img_rank	object_by	object_width	object_height	object_area	object_mean	object_major	object_minor	object_feret	object_area_exc	object_thickr	object_esd	object_elongation	object_range	object_meanpos	object_centroids	object_cv	object_sr	object_perimareaexc	object_feretareaexc	object_perimferet	object_perimmajor	object_circex	object_cdexc	object_kurt_mean	object_skew_mean	object_convperim_perim	object_convarea_area	object_symetrieh_area	object_symetriev_area	object_nb1_area	object_nb2_area	object_nb3_area	object_nb1_range	object_nb2_range	object_nb3_range	object_median_mean	object_median_mean_range	object_skeleton_area	sample_id	sample_dataportal_descriptor	sample_project	sample_ship	sample_filename	sample_samplinggear	sample_volconc	sample_comment	sample_tot_vol	process_id	process_software	process_start_date	process_start_time	process_analysed_color_channel	process_background_method	process_nb_images	process_rolling	process_pixel	process_upper	process_gamma_value	process_vignette_color	process_esd_min	process_esd_max	process_lut_offset	process_lut_slope	process_remove_objects_on_sides	process_stop_n_images	process_scale	process_stop_after_m_objects	process_nb_of_rawfile_images_in_folder	process_grey_auto_adjust	acq_id	acq_instrument	acq_cell_type	acq_magnification	acq_sub_part
+[t]	[t]	[f]	[f]	[t]	[t]	[t]	[f]	[f]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[f]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[f]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[t]	[f]
+m106_mn01_n3_sml_1120	Extra3	18.0	-24.4166666666667	20140421	420		300.0	600.0	predicted	Udale Crank	ucrank4@gov.uk	20151111	153120	m106_mn01_n3_sml_1111.jpg		10961.0	25.0	38.0	499.9	205.2	28.0	21	29	0.0	2.0	2.01																												m106_mn01_n3_sml		2014_m106_sml	meteor	m106_mn01_n3_sml	net	9.75	xx		zooprocess_m106_mn01_n3_sml	zooprocess_pid_to_ecotaxa_7.23_2017/03/10	20150829	2256		rolling_ball	2	20	10.6	245	2	b&w_no_background	200	1000	0	1	yes		1		2		generic_m106_mn01_n3_sml	epson	200-500	1
