+++ QA/py/tests/test_admin.py
@@ -35,6 +35,15 @@ def test_admin_images(database, fastapi, caplog):
     assert rsp.json() == "Digest for 0 images done."
 
 
+def do_nightly(fastapi):
+    rsp = fastapi.get(NIGHTLY_URL, headers=ADMIN_AUTH)
+    assert rsp.status_code == status.HTTP_200_OK
+
+    job_id = rsp.json()
+    job = wait_for_stable(job_id)
+    check_job_ok(job)
+
+
 def test_nightly_job(database, fastapi, caplog, tstlogs):
     # TODO: Not a real test, as we can't know in advance when the test runs, so the output
     # can't be verified against a reference.
@@ -56,22 +65,14 @@ def test_nightly_job(database, fastapi, caplog, tstlogs):
     # Only Admin can
     caplog.set_level(logging.DEBUG)
 
-    def do_nightly():
-        rsp = fastapi.get(NIGHTLY_URL, headers=ADMIN_AUTH)
-        assert rsp.status_code == status.HTTP_200_OK
-
-        job_id = rsp.json()
-        job = wait_for_stable(job_id)
-        check_job_ok(job)
-
-    do_nightly()
+    do_nightly(fastapi)
     msgs = len(
         [msg for msg in caplog.messages if msg.startswith("About to clean 3 jobs")]
     )
     assert msgs > 0
 
     # Second cleanup must do nothing
-    do_nightly()
+    do_nightly(fastapi)
     msgs = len(
         [msg for msg in caplog.messages if msg.startswith("About to clean 0 jobs")]
     )
